finish award rule infra test

+++ ddd-main-domain/src/main/java/org/example/domain/strategy/repository/IStrategyRepo.java
@@ -22,5 +22,5 @@ public interface IStrategyRepo {
 
     StrategyRuleEntity queryStrategyRuleByIdAndName(Long strategyId, String ruleName);
 
-    AwardRuleEntity queryAwardRuleByIdAndName(Long strategyId, Long awardId, String ruleName);
+    AwardRuleEntity queryAwardRuleByIdAndName(Long strategyId, Integer awardId, String ruleName);
 }
